requirements.txt
cairocffi==1.3.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
cssselect2==0.5.0
defusedxml==0.7.1
dominate==2.6.0
Flask~=2.1.2
Flask-Bootstrap==3.3.7.1
gunicorn==20.1.0
hashids==1.3.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
Pillow~=9.1.1
pycparser==2.21
pyTelegramBotAPI==4.4.0
qrcode==7.3.1
requests==2.27.1
tinycss2==1.1.1
urllib3==1.26.9
visitor==0.1.3
webencodings==0.5.1
Werkzeug==2.1.1
zipp==3.8.0

svgutils~=0.3.4